refactor test into separate files [Finishes #148445763]
